// Frontend_Complete/src/app/models/book-slot.model.ts
export class BookSlotRegister{

    id:number=0;
    employeeId: string ='';
    employeeName: string ='';
    department: string ='';
    seats: number =0;
    floor: string ='';
    roomNumber: string ='';
    dateTime: Date = new Date();
    startTime: string ='';
    endTime: string ='';

    
}